Modified   Common.h Modified   DSPs/AutoWah.h Modified   DSPs/Chorus.h Modified   DSPs/Compressor.h Modified   DSPs/Distortion.h Modified   DSPs/Flanger.h Modified   DSPs/IDSP.h Modified   DSPs/Looper.h Modified   DSPs/NoiseGate.h Modified   DSPs/Overdrive.h Modified   DSPs/Phaser.h Modified   DSPs/Reverb.h Modified   DSPs/Sustain.h Modified   DSPs/Test.h Modified   DSPs/Tremolo.h Modified   DSPs/Wah.h Modified   Debug.h Modified   Filters/BandPassFilter.h Modified   Filters/BandStopFilter.h Modified   Filters/Filter.h Modified   PersistentBlob.h

+++ DSPs/IDSP.h
@@ -10,7 +10,7 @@ class IDSP
 	static_assert(ARE_TYPES_THE_SAME(T, float) || ARE_TYPES_THE_SAME(T, double), "T must be float or double");
 
 public:
-	virtual void ProcessBuffer(T *Buffer, uint16 Count) = 0;
+	virtual void ProcessBuffer(T *Buffer, uint8 Count) = 0;
 };
 
 #endif
