+++ DSPs/Wah.h
@@ -69,7 +69,7 @@ public:
 		return m_Ratio;
 	}
 
-	void ProcessBuffer(T *Buffer, uint16 Count) override
+	void ProcessBuffer(T *Buffer, uint8 Count) override
 	{
 		for (uint16 i = 0; i < Count; ++i)
 			Buffer[i] = Process(Buffer[i]);
