+++ Filters/Filter.h
@@ -10,7 +10,7 @@ class Filter
 public:
 	virtual T Process(T Value) = 0;
 
-	virtual void ProcessBuffer(T *Buffer, uint16 Count)
+	virtual void ProcessBuffer(T *Buffer, uint8 Count)
 	{
 		for (uint16 i = 0; i < Count; ++i)
 			Buffer[i] = Process(Buffer[i]);
